fix: crud de historico e integracao front

+++ backend/src/main/java/fitt_nutri/example/demo/model/AnthropometricDataModel.java
@@ -34,7 +34,7 @@ public class AnthropometricDataModel {
 
     private Double gorduraVisceral;
 
-    private Integer taxaMetabolicaBasal;
+    private Double taxaMetabolicaBasal;
 
     private Integer idadeMetabolica;
